Add a CTRL-S key binding, delegating to ProseMirror.prototype.save
This commit introduces a simple API for adding a `save` feature to
ProseMirror. The key binding is already put for a `save` command.
The latter delegated to `#save` on the editor, if any. Adding such
a saving function can simply be made by assigning a function to
`ProseMirror.prototype.save`.